erofs-utils: mkfs: introduce multi-threaded compression for non-dedupe scenarios
Currently, the inefficiency of EROFS compressed image creation is a much criticized
problem, and this patch attempts to address it by running the compression algorithm
in parallel in a inner-file manner.

Specifically, each input file larger than 16MB is splited into segments,
and each worker thread compresses a segment as if it were a separate
file. Finally, the main thread merges all the compressed segments.

Multi-threaded compression is not compatible with -Ededupe now.

Signed-off-by: Yifan Zhao <zhaoyifan@sjtu.edu.cn>